Add logic to substrings function
It has two parameters, one is a word or a string with many words and
the second is a "dictionary" or an array of substrings.

The objective is to create a hash that contains all the dictionary
substrings found in the word(s) as keys and the count of each one
as values.

To do this, first it creates an array of the word(s), then convert them
to lowercase because it is irrelevant to match the case in this project.

Then create a new array that will store the arrays of the substrings
found in the words. For that it will iterate each word first and then
for each iteration there is another one to iterate each element of the
dictionary to see if an element is substring of the current word; if it
is then it pushes it to the current array and when it finishes it pushes
that array to the other one.

When that process has finished, the main arary is now an array of
arrays; to create the hash first it is needed to flatten it. Now done
that, there is another iteration with the reduce method where it creates
a new entry in the hash when a new substring is found, in case it
already exists, just adds one to the value of that element. And when
this loop ends it returns the hash.